ENH: Profile improf loading, add --mem/--show options
Allow users to skip memory collection to avoid loading psutil and its
dependencies. There's no avoiding sys and time.

Use --no-mem as the first flag to avoid memory and --mem to turn it on
later in the list.

Use --show after an import to show the modules loaded, and --no-show as
the last argument to avoid showing the final set of modules loaded.